chore: refactor integrations
