+++ app/integrations/postgresql.py
@@ -13,6 +13,7 @@ from typing import Any
 
 from pydantic import Field, field_validator
 
+from app.integrations._relational import env_int, env_str, resolve_stored_or_env_config
 from app.strict_config import StrictConfigModel
 
 DEFAULT_POSTGRESQL_PORT = 5432
@@ -77,20 +78,18 @@ def build_postgresql_config(raw: dict[str, Any] | None) -> PostgreSQLConfig:
 
 def postgresql_config_from_env() -> PostgreSQLConfig | None:
     """Load a PostgreSQL config from env vars."""
-    host = os.getenv("POSTGRESQL_HOST", "").strip()
-    database = os.getenv("POSTGRESQL_DATABASE", "").strip()
+    host = env_str("POSTGRESQL_HOST")
+    database = env_str("POSTGRESQL_DATABASE")
     if not host or not database:
         return None
     return build_postgresql_config(
         {
             "host": host,
-            "port": int(_pg_port)
-            if (_pg_port := os.getenv("POSTGRESQL_PORT", "").strip()).isdigit()
-            else DEFAULT_POSTGRESQL_PORT,
+            "port": env_int("POSTGRESQL_PORT", DEFAULT_POSTGRESQL_PORT),
             "database": database,
-            "username": os.getenv("POSTGRESQL_USERNAME", DEFAULT_POSTGRESQL_USER).strip(),
+            "username": env_str("POSTGRESQL_USERNAME", DEFAULT_POSTGRESQL_USER),
             "password": os.getenv("POSTGRESQL_PASSWORD", ""),
-            "ssl_mode": os.getenv("POSTGRESQL_SSL_MODE", DEFAULT_POSTGRESQL_SSL_MODE).strip(),
+            "ssl_mode": env_str("POSTGRESQL_SSL_MODE", DEFAULT_POSTGRESQL_SSL_MODE),
         }
     )
 
@@ -104,36 +103,24 @@ def resolve_postgresql_config(
     Credentials (username, password, ssl_mode) are resolved from the stored
     integration or environment variables so they never appear in tool signatures.
     """
-    from app.integrations.store import get_integration
-
-    stored = get_integration("postgresql")
-    if stored:
-        creds = stored.get("credentials", {})
-        return build_postgresql_config(
-            {
-                "host": host,
-                "port": creds.get("port", port),
-                "database": database,
-                "username": creds.get("username", DEFAULT_POSTGRESQL_USER),
-                "password": creds.get("password", ""),
-                "ssl_mode": creds.get("ssl_mode", DEFAULT_POSTGRESQL_SSL_MODE),
-            }
-        )
-
-    env_cfg = postgresql_config_from_env()
-    if env_cfg:
-        return build_postgresql_config(
-            {
-                "host": host,
-                "port": port,
-                "database": database,
-                "username": env_cfg.username,
-                "password": env_cfg.password,
-                "ssl_mode": env_cfg.ssl_mode,
-            }
-        )
-
-    return build_postgresql_config({"host": host, "port": port, "database": database})
+    return resolve_stored_or_env_config(
+        "postgresql",
+        host=host,
+        database=database,
+        port=port,
+        build_config=build_postgresql_config,
+        env_loader=postgresql_config_from_env,
+        extra_from_credentials=lambda credentials: {
+            "username": credentials.get("username", DEFAULT_POSTGRESQL_USER),
+            "password": credentials.get("password", ""),
+            "ssl_mode": credentials.get("ssl_mode", DEFAULT_POSTGRESQL_SSL_MODE),
+        },
+        extra_from_env=lambda config: {
+            "username": config.username,
+            "password": config.password,
+            "ssl_mode": config.ssl_mode,
+        },
+    )
 
 
 def _get_connection(config: PostgreSQLConfig) -> Any:
